Alterações para reset de senha de forma correta

+++ Users/account/helpers.py
@@ -1,5 +1,6 @@
 from Users.account import *
 
+from AppCore.core.exceptions.exceptions import SystemErrorException
 from AppCore.core.helpers.helpers import ModelInstanceHelpers
 
 from .models import EmailAccountCode
@@ -11,14 +12,21 @@ class AccountHelper(ModelInstanceHelpers):
         
         return User.objects.filter(email=email, profiles__type=type_profile).exists()
 
-    def del_codes_expired(self, email):
-        EmailAccountCode.objects.filter(
-            Q(
-                created_at__lt=timezone.now() - timezone.timedelta(minutes=30)
-            ) | Q(
-                email=email
-            )
-        ).delete()
+    def del_codes_expired(self, email=None):
+        if self.object_instance or email:
+            if not email:
+                email = self.object_instance.email
+
+            EmailAccountCode.objects.filter(
+                Q(
+                    created_at__lt=timezone.now() - timezone.timedelta(minutes=30)
+                ) | Q(
+                    email=email
+                )
+            ).delete()
+        
+        else:
+            raise SystemErrorException('Parâmetros insuficientes para deletar códigos expirados.')
         
     def validate_valid_code(self, email, code):
         try:
